+++ packages/libs/core/src/images/imageOptimizer.ts
@@ -75,6 +75,23 @@ export function getMaxAge(str: string | undefined): number {
   return minimum;
 }
 
+/**
+ * If Basepath set, it needs to be removed from URL
+ *
+ * Not normalised -> error 403
+ * url: '<base-path>/assets/images/logo.svg',
+ *
+ * Normalised -> 200
+ * url: '/assets/images/logo.svg',
+ */
+export function normaliseUri(uri: string, basePath: string): string {
+  if (uri.startsWith(basePath)) {
+    uri = uri.slice(basePath.length);
+  }
+
+  return uri;
+}
+
 export async function imageOptimizer(
   basePath: string,
   imagesManifest: ImagesManifest | undefined,
@@ -116,7 +133,14 @@ export async function imageOptimizer(
   let isAbsolute: boolean;
 
   if (url.startsWith("/")) {
-    href = url;
+    // Ensure that Basepath is in the URL, otherwise, a 400 is triggered (same behaviour as Nextjs)
+    if (basePath !== "/" && !url.startsWith(basePath)) {
+      res.statusCode = 400;
+      res.end('"Basepath" set but not added to the URL');
+      return { finished: true };
+    }
+
+    href = normaliseUri(url, basePath);
     isAbsolute = false;
   } else {
     let hrefParsed: URL;
@@ -236,11 +260,9 @@ export async function imageOptimizer(
   } else {
     let objectKey;
     try {
-      if (
-        href.startsWith(`${basePath}/static`) ||
-        href.startsWith(`${basePath}/_next/static`)
-      ) {
-        objectKey = href; // static files' URL map to the key directly e.g /static/ -> static
+      // note: basepath is already removed by URI normalization above
+      if (href.startsWith(`/static`) || href.startsWith(`/_next/static`)) {
+        objectKey = `${basePath}${href}`; // static files' URL map to the key prefixed with basepath e.g /static/ -> static
       } else {
         objectKey = `${basePath}/public` + href; // public file URLs map from /public.png -> public/public.png
       }