+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.4-alpha.0](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3...v3.5.4-alpha.0) (2021-11-10)
+
+### Bug Fixes
+
+- **core:** fix image optimizer for images with basepath in /\_next/static or /static/ ([#2033](https://github.com/serverless-nextjs/serverless-next.js/issues/2033)) ([163c95a](https://github.com/serverless-nextjs/serverless-next.js/commit/163c95a6556b7bf0a58eca419b292fc1a612510c))
+
+## [3.5.3](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3-alpha.6...v3.5.3) (2021-11-09)
+
+**Note:** Version bump only for package serverless-nextjs-monorepo
+
+## [3.5.3-alpha.6](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3-alpha.5...v3.5.3-alpha.6) (2021-11-09)
+
+### Bug Fixes
+
+- **core:** add url normalization to image optimizer URL ([#2015](https://github.com/serverless-nextjs/serverless-next.js/issues/2015)) ([5c4f841](https://github.com/serverless-nextjs/serverless-next.js/commit/5c4f8411268e4b874fe296bd2ac5b880bd9addc1))
+- **core:** render static pages for all http methods, for options methods return allowed methods ([#2024](https://github.com/serverless-nextjs/serverless-next.js/issues/2024)) ([9fe3216](https://github.com/serverless-nextjs/serverless-next.js/commit/9fe32167064f5a220ba1d55d8400673286752efc))
+
 ## [3.5.3-alpha.5](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3-alpha.4...v3.5.3-alpha.5) (2021-11-08)
 
 ### Bug Fixes